requirements.txt
apache-airflow
pandas
requests
tqdm
nba_api
sqlalchemy
psycopg2-binary
SQLAlchemy